+++ test/commands.js
@@ -406,6 +406,7 @@ test('[commands] help', t => {
   --skip-domain-reachability-check  Do not run pre-flight check for domain reachability.
   --access-log-errors-only            Display only errors in the access log (HTTP status codes _4xx_ and _5xx_).
   --access-log-disable                Completely disable the access log. Not even errors are logged.
+  --allow-embeds                      Allow site contents to be embedded on other sites via iframes.
 
   For serve command:
 
@@ -515,6 +516,7 @@ test('[commands] help', t => {
     --skip-domain-reachability-check  Do not run pre-flight check for domain reachability.
     --access-log-errors-only            Display only errors in the access log (HTTP status codes _4xx_ and _5xx_).
     --access-log-disable                Completely disable the access log. Not even errors are logged.
+    --allow-embeds                      Allow site contents to be embedded on other sites via iframes.
 
     --sync-to                   The host to sync to (other sync options only relevant if this is supplied).
     --sync-from                 The folder to sync from.
@@ -606,6 +608,7 @@ test('[commands] help', t => {
     --skip-domain-reachability-check  Do not run pre-flight check for domain reachability.
     --access-log-errors-only            Display only errors in the access log (HTTP status codes _4xx_ and _5xx_).
     --access-log-disable                Completely disable the access log. Not even errors are logged.
+    --allow-embeds                      Allow site contents to be embedded on other sites via iframes.
 
     --sync-to                   The host to sync to (other sync options only relevant if this is supplied).
     --sync-from                 The folder to sync from.
@@ -694,6 +697,7 @@ test('[commands] help', t => {
     --skip-domain-reachability-check  Do not run pre-flight check for domain reachability.
     --access-log-errors-only            Display only errors in the access log (HTTP status codes _4xx_ and _5xx_).
     --access-log-disable                Completely disable the access log. Not even errors are logged.
+    --allow-embeds                      Allow site contents to be embedded on other sites via iframes.
 
     --sync-to                   The host to sync to (other sync options only relevant if this is supplied).
     --sync-from                 The folder to sync from.